ai_service/requirements.txt
pika==1.3.1
boto3==1.28.44
openai==0.28.0
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0 